chore(sidecar): commitments server refactor
